separate IOV into components, add logging
