+++ src/iov/player.js
@@ -1,179 +1,7 @@
 import Debug from 'debug';
-import uuidv4 from 'uuid/v4';
 
-const DEBUG_PREFIX = 'clsp:iov:';
-const debug = Debug(`${DEBUG_PREFIX}:main`);
-
-/**
- * Internet of Video client. This module uses the MediaSource API to
- * deliver video content streamed through MQTT from distributed sources.
- */
-
-function mediaSourceCheck () {
-  // For the MAC
-  var NoMediaSourceAlert = false;
-
-  window.MediaSource = window.MediaSource || window.WebKitMediaSource;
-
-  if (!window.MediaSource){
-    if (NoMediaSourceAlert === false) {
-      alert("Media Source Extensions not supported in your browser"+
-         ": Claris Live Streaming will not work!");
-    }
-
-    NoMediaSourceAlert = true;
-  }
-}
-
-class MqttConduitCollection {
-  constructor () {
-    this._conduits = {};
-  }
-
-  set (id, conduit) {
-    this._conduits[id] = conduit;
-  }
-
-  addFromIov (iov) {
-    this.set(iov.config.clientId, mqttConduit(iov.config, () => {
-      iov.config.appStart(iov);
-
-      // the mse service will stop streaming to us if we don't send
-      // a message to iov/stats within 1 minute.
-      iov._statsTimer = setInterval(() => {
-        iov.statsMsg.inkbps = (iov.statsMsg.byteCount * 8) / 30000.0;
-        iov.statsMsg.byteCount = 0;
-
-        self.publish("iov/stats",iov.statsMsg);
-
-        debug('iov status', iov.statsMsg);
-      }, 5000);
-    }));
-  }
-
-  getById (id) {
-    return this._conduits[id];
-  }
-
-  exists (id) {
-    return this._conduits.hasOwnProperty(id);
-  }
-}
-
-class MqttTransport {
-  constructor (iov) {
-    this.debug = Debug(`${DEBUG_PREFIX}:MqttTransport`);
-
-    this.debug('constructing...');
-
-    // setup stats
-    iov.statsMsg = {
-      byteCount: 0,
-      inkbps: 0,
-      host: document.location.host,
-      clientId: iov.config.clientId
-    };
-
-    this.clientId = iov.config.clientId;
-    this.iov = iov;
-
-    this.conduit = iov.mqttConduitCollection.addFromIov(iov);
-  }
-
-  // create a temporary resp_topic to receive a query response
-  // upon response remove the temporary topic. Assume both request
-  // and response are json formateed text.
-  transaction (topic, callback, data) {
-    this.debug('transaction...');
-
-    this.conduit.transaction(topic, (event) => {
-      this.debug('transaction callback...', event);
-      if (callback) {
-        callback(event);
-      }
-    }, data);
-  }
-
-  subscribe (topic, callback) {
-    this.debug('subscribe...');
-
-    this.conduit.subscribe(topic, (event) => {
-      this.debug('subscribe callback...', event);
-      if (callback) {
-        callback(event);
-      }
-    });
-  }
-
-  unsubscribe (topic, callback) {
-    this.debug('unsubscribe...');
-
-    this.conduit.unsubscribe(topic, (event) => {
-      this.debug('unsubscribe callback...', event);
-      if (callback) {
-        callback(event);
-      }
-    });
-  }
-
-  publish (topic, data, callback) {
-    this.debug('publish...');
-
-    this.conduit.publish(topic, data, (event) => {
-      this.debug('publish callback...', event);
-      if (callback) {
-        callback(event);
-      }
-    });
-  }
-}
-
-/**
- * route inbound messages/data to player's event handlers.
- */
-class MqttTopicHandlers {
-  constructor (iov) {
-    this.iov = iov;
-    this._handlers = {};
-  }
-
-  get (topic) {
-    return this._handlers[topic];
-  }
-
-  register (topic, callback) {
-    this._handlers[topic] = callback;
-  }
-
-  unregister (topic) {
-    if (this.exists(topic)) {
-      delete this._handlers[topic];
-    }
-  }
-
-  exists (topic) {
-    return this._handlers.hasOwnProperty(topic);
-  }
-
-  // central entry point for all MQTT inbound traffic.
-  msghandler (message) {
-    var topic = message.destinationName;
-
-    if (!this.exists(topic)) {
-      console.log("No handler for " + topic);
-      console.log("message dropped");
-      console.log(message);
-
-      return;
-    }
-
-    try {
-      this.get(topic)(message);
-    } catch (e) {
-      this.iov.events.exception(`${topic} handler exception`, e);
-    }
-  }
-}
+const DEBUG_PREFIX = 'clsp:iov';
+const debug = Debug(`${DEBUG_PREFIX}:player`);
 
 /**
  * Responsible for receiving stream input and routing it to the media source
@@ -183,7 +11,7 @@ class MqttTopicHandlers {
  * var player = iov.player();
  * player.play( video_element_id, stream_name );
 */
-var _player = function(iov){
+export default function (iov) {
     var self = {};
 
     /*
@@ -512,87 +340,3 @@ var _player = function(iov){
 
     return self;
 };
-
-export default class IOV {
-  constructor (config) {
-    this.config = {
-      // web socket address defaults to the address of the server that loaded this page.
-      wsbroker: config.address,
-      // default port number
-      wsport: config.port,
-      // default clientId
-      clientId: uuidv4(),
-      // to be overriden by user.
-      appStart: config.appStart,
-      useSSL: config.useSSL || false,
-      videoElement: config.videoElement
-    };
-
-    // handle inbound messages from MQTT, including video
-    // and distributes them to players.
-    this.mqttTopicHandlers = new MqttTopicHandlers(this);
-
-    this.mqttConduitCollection = new MqttConduitCollection();
-    this.transport = new MqttTransport(this);
-
-    this.events = {
-      connection_lost : function(responseObject) {
-        //TODO close all players and display an error message
-        console.log("MQTT connection lost");
-        console.log(responseObject);
-      },
-
-      on_message: this.mqttTopicHandlers.msghandler,
-
-      // generic exception handler
-      exception: function(text,e) {
-        console.log(text);
-        if (typeof e !== 'undefined') {
-          console.log(e.stack);
-        }
-      }
-    };
-  }
-
-  initialize () {
-    mediaSourceCheck();
-
-    // route inbound data from a frame running mqtt to the appropriate player
-    window.addEventListener('message', (event) => {
-      debug('message received', event.data)
-
-      var clientId = event.data.clientId;
-
-      if (!this.mqttConduitCollection.exists(clientId)) {
-        return;
-      }
-
-      var conduit = this.mqttConduitCollection.get(clientId);
-      var eventType = event.data.event;
-
-      switch (eventType) {
-        case 'data':
-          conduit.inboundHandler(event.data);
-          break;
-        case 'ready':
-          conduit.onReady();
-          break;
-        case 'fail':
-          console.error('Failure Message!');
-          console.error(event);
-          break;
-        default:
-          console.log("No match for event = " + eventType);
-      }
-    });
-  }
-
-  player () {
-    return _player(this);
-  }
-
-  // query remote server and get a list of all stream names
-  getAvailableStreams (callback) {
-    this.transport.transaction('iov/video/list', callback, {});
-  }
-}
